refactor(research): refactor imports and improve random key usage in evaluation
- Updated import statements in performance_optimization.py and quantum_optimization.py to import FederatedGraphRL from core.federated instead of core.types.
- Modified quick_evaluation method in autonomous_meta_learning.py to use jax.random.PRNGKey and jax.random.split for reproducible random number generation instead of jnp.random.normal.

These changes improve code clarity and ensure consistent random number generation for evaluations.
