+++ pg_neo_graph_rl/research/quantum_optimization.py
@@ -24,7 +24,7 @@ import networkx as nx
 import numpy as np
 import optax
 
-from ..core.types import FederatedGraphRL
+from ..core.federated import FederatedGraphRL
 
 
 class QuantumState(NamedTuple):
